% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_particle_slip_values.R
\name{get_particle_slip_values}
\alias{get_particle_slip_values}
\title{get_particle_slip_values - Retrieve Slip Values from Model Experiment}
\usage{
get_particle_slip_values(topo, particles)
}
\arguments{
\item{particles}{}
}
\value{

}
\description{
get_particle_slip_values - Retrieve Slip Values from Model Experiment
}
\examples{


}
