% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_resampled_topo.R
\name{get_resampled_topo}
\alias{get_resampled_topo}
\title{get_resampled_topo - Get the Resampled of the Input Topography}
\usage{
get_resampled_topo(topo, crs, resolution = c(0.0041656, 0.0041656), ...)
}
\arguments{
\item{...}{}
}
\value{
bathy
}
\description{
get_resampled_topo - Get the Resampled of the Input Topography
}
