% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_topo_projection.R
\name{get_topo_projection}
\alias{get_topo_projection}
\title{get_topo_projection - Get the Raster Projection of the Input Topography}
\usage{
get_topo_projection(topo, crs, resolution = c(0.0041656, 0.0041656), ...)
}
\arguments{
\item{...}{}
}
\value{
raster
}
\description{
get_topo_projection - Get the Raster Projection of the Input Topography
}
