% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LandslideR-package.R
\docType{package}
\name{LandslideR-package}
\alias{LandslideR}
\alias{LandslideR-package}
\title{LandslideR: LandslideR: Ultradiscrete Cellular Automata Model for Generalized Mass Failure}
\description{
The LandslideR is a tool for modelling mass failures such as subaerial and submarine landslides, that is based on Ultradiscretized Cellular Automata derived from two-dimensional diffusion equation.
}
\keyword{internal}
