% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nwbt.slide.area}
\alias{nwbt.slide.area}
\alias{nwbt.slide.area.scarp}
\title{Northwest Borneo Datasets}
\format{
An object of class \code{bathy} with 216 rows and 240 columns.

An object of class \code{bathy} with 119 rows and 97 columns.
}
\usage{
nwbt.slide.area

nwbt.slide.area.scarp
}
\description{
These datasets contain the area of concern where there could be
submarine landslide in the tip the Northwest Borneo and the
assumed scarp area
}
\keyword{datasets}
