% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replaceZero.R
\name{replaceZero}
\alias{replaceZero}
\title{replaceZero - Replace Zeroes with NA}
\usage{
replaceZero(x)
}
\arguments{
\item{x}{}
}
\value{

}
\description{
replaceZero - Replace Zeroes with NA
}
\examples{

}
