% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_slide_to_particles.R
\name{set_slide_to_particles}
\alias{set_slide_to_particles}
\title{set_slide_to_particles}
\usage{
set_slide_to_particles(topo, slide_topo, slip_value = 1000)
}
\arguments{
\item{slip_value}{}
}
\value{

}
\description{
set_slide_to_particles
}
\examples{

}
