% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UltradiscreteCA.R
\name{UltradiscreteCA}
\alias{UltradiscreteCA}
\title{Ultradiscretized Cellular Automata Model for Landslides}
\description{
'UltradiscreteCA' class implements the Ultradiscretized Cellular Automata
  rules derived from two dimensional diffusion equation
}
\details{
R6 Class representing a
}
\examples{
\dontrun{
data("nwbt.slide.area")
data("nwbt.slide.area.scarp")

# Model Initializiation
model <- UltradiscreteCA$new(
domain = nwbt.slide.area,          # ASCII/NetCDF Format
slip_area = nwbt.slide.area.scarp, # Assumed location of the landslide
rock_size = 0.5 # unit in meters
)
}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{UltradiscreteCA$new()}}
\item \href{#method-get_iteration}{\code{UltradiscreteCA$get_iteration()}}
\item \href{#method-simulate}{\code{UltradiscreteCA$simulate()}}
\item \href{#method-get_simulation_particles}{\code{UltradiscreteCA$get_simulation_particles()}}
\item \href{#method-get_simulation_slip_values}{\code{UltradiscreteCA$get_simulation_slip_values()}}
\item \href{#method-get_simulation_displacements}{\code{UltradiscreteCA$get_simulation_displacements()}}
\item \href{#method-save_plots}{\code{UltradiscreteCA$save_plots()}}
\item \href{#method-save_output}{\code{UltradiscreteCA$save_output()}}
\item \href{#method-get_particles}{\code{UltradiscreteCA$get_particles()}}
\item \href{#method-get_slip_values}{\code{UltradiscreteCA$get_slip_values()}}
\item \href{#method-clone}{\code{UltradiscreteCA$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UltradiscreteCA$new(domain, slip_area, rock_size = 0.5, scarp_depth = 1000)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_iteration"></a>}}
\if{latex}{\out{\hypertarget{method-get_iteration}{}}}
\subsection{Method \code{get_iteration()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UltradiscreteCA$get_iteration(particle)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-simulate"></a>}}
\if{latex}{\out{\hypertarget{method-simulate}{}}}
\subsection{Method \code{simulate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UltradiscreteCA$simulate(iterations = 5, save = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_simulation_particles"></a>}}
\if{latex}{\out{\hypertarget{method-get_simulation_particles}{}}}
\subsection{Method \code{get_simulation_particles()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UltradiscreteCA$get_simulation_particles()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_simulation_slip_values"></a>}}
\if{latex}{\out{\hypertarget{method-get_simulation_slip_values}{}}}
\subsection{Method \code{get_simulation_slip_values()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UltradiscreteCA$get_simulation_slip_values()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_simulation_displacements"></a>}}
\if{latex}{\out{\hypertarget{method-get_simulation_displacements}{}}}
\subsection{Method \code{get_simulation_displacements()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UltradiscreteCA$get_simulation_displacements(every_n = 5)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-save_plots"></a>}}
\if{latex}{\out{\hypertarget{method-save_plots}{}}}
\subsection{Method \code{save_plots()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UltradiscreteCA$save_plots(
  type = "slide",
  every_n = 1,
  save_location = getwd(),
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-save_output"></a>}}
\if{latex}{\out{\hypertarget{method-save_output}{}}}
\subsection{Method \code{save_output()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UltradiscreteCA$save_output(
  every_n = 5,
  file_name_prefix = "slide",
  folder_location = getwd(),
  crs = "+proj=longlat +datum=WGS84 +ellps=WGS84 +towgs84=0,0,0",
  resolution = c(0.0041656, 0.0041656)
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_particles"></a>}}
\if{latex}{\out{\hypertarget{method-get_particles}{}}}
\subsection{Method \code{get_particles()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UltradiscreteCA$get_particles()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_slip_values"></a>}}
\if{latex}{\out{\hypertarget{method-get_slip_values}{}}}
\subsection{Method \code{get_slip_values()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UltradiscreteCA$get_slip_values()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UltradiscreteCA$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
